refactor(frontend): remove duplicated style and source helpers
