+++ src/views/HomeView/utils/projectSource.ts
@@ -1,36 +1,9 @@
 import type { LocalProject } from '~/constants/localProject'
 import { ProjectKind } from '~/constants/localProject'
 import {
-  normalizeProjectSourceExternalUrl,
-  sourceNameFromUrl,
+  projectSourceName,
+  projectSourceUrl,
 } from '~/utils/projectSource'
-
-export function isSourceProject(project: LocalProject) {
-  return project.kind === ProjectKind.FORK || project.kind === ProjectKind.CLONE
-}
-
-export function projectSourceUrl(project: LocalProject) {
-  return isSourceProject(project) ? project.fromUrl?.trim() || '' : ''
-}
-
-export function projectSourceExternalUrl(project: LocalProject) {
-  const sourceUrl = projectSourceUrl(project)
-  if (!sourceUrl)
-    return ''
-
-  return normalizeProjectSourceExternalUrl(sourceUrl)
-}
-
-export function projectSourceName(project: LocalProject) {
-  const sourceUrl = projectSourceUrl(project)
-  return isSourceProject(project)
-    ? project.fromName?.trim() || (sourceUrl ? sourceNameFromUrl(sourceUrl) : '')
-    : ''
-}
-
-export function hasProjectSource(project: LocalProject) {
-  return !!projectSourceName(project)
-}
 
 export function projectSourcePrefix(project: LocalProject, t: (key: string) => string) {
   return project.kind === ProjectKind.FORK
